safari-600.7-branch/LayoutTests/platform/mac/accessibility/spinbutton-valuedescription.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body id="body">

<div role="spinbutton" id="spin" aria-valuetext="Monday" aria-valuemin="1" aria-valuemax="7" aria-valuenow="4">spinbutton</div>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that aria-value* properties work on spinbutton roles.");

    if (window.accessibilityController) {

          var spin = accessibilityController.accessibleElementById("spin");
          shouldBe("spin.valueDescription", "'AXValueDescription: Monday'");
          shouldBe("spin.stringValue", "'AXValue: 4'");
          shouldBe("spin.minValue", "1");
          shouldBe("spin.maxValue", "7");

          shouldBeTrue("spin.isAttributeSupported('AXValueDescription')");
          shouldBeTrue("spin.isAttributeSupported('AXMaxValue')");
          shouldBeTrue("spin.isAttributeSupported('AXMinValue')");
    }

</script>

<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
