LayoutTests/platform/mac/accessibility/spinbutton-valuedescription-expected.txt

spinbutton
This tests that aria-value* properties work on spinbutton roles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS spin.valueDescription is 'Monday'
PASS spin.stringValue is 'AXValue: 4'
PASS spin.minValue is 1
PASS spin.maxValue is 7
PASS spin.isAttributeSupported('AXValueDescription') is true
PASS spin.isAttributeSupported('AXMaxValue') is true
PASS spin.isAttributeSupported('AXMinValue') is true
PASS successfullyParsed is true

TEST COMPLETE
